Fixes for Windows.

+++ tools/mbrpad.c
@@ -12,8 +12,8 @@ int main(int argc, char* argv[]) {
     int fd = open(filename, O_RDWR);
     off_t length = lseek(fd, 0, SEEK_END);
     if (length > 510) {
-        fprintf(stderr, "file %s is larger than 510 bytes (size: %ju)\n",
-                filename, (uintmax_t)length);
+        fprintf(stderr, "file %s is larger than 510 bytes (size: %llu)\n",
+                filename, (unsigned long long)length);
         return 1;
     }
     lseek(fd, 510, SEEK_SET);
